pyproject.toml
[tool.poetry]
name = "artefact"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.1.3"
django-channels = "^0.7.0"
django-htmx = "^1.21.0"
httpx = "^0.27.2"
django-polymorphic = "^3.1.0"
django-livereload-server = "^0.5.1"
colorlog = "^6.9.0"
python-dotenv = "^1.0.1"
daphne = "^4.1.2"
uuid = "^1.30"
openai = "^1.55.2"
ipdb = "^0.13.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
